refactor: Simplified tests (create-options)
