+++ tests/migration/ember-app/steps/create-options/javascript.test.ts
@@ -1,23 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
 import { createOptions } from '../../../../../src/migration/ember-app/steps/index.js';
+import { inputProject } from '../../../../fixtures/ember-app/javascript/index.js';
 import {
   codemodOptions,
   options,
 } from '../../../../helpers/shared-test-setups/ember-app/javascript.js';
 
 test('migration | ember-app | steps | create-options > javascript', function () {
-  const inputProject = {
-    'package.json': JSON.stringify(
-      {
-        name: '@ijlee2/ember-workshop-app',
-        version: '0.0.0',
-      },
-      null,
-      2,
-    ),
-  };
-
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(createOptions(codemodOptions), options);
