+++ tests/migration/ember-app/steps/create-options/sass.test.ts
@@ -1,13 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
 import { createOptions } from '../../../../../src/migration/ember-app/steps/index.js';
-import { inputProject } from '../../../../fixtures/ember-app/typescript/index.js';
+import { inputProject } from '../../../../fixtures/ember-app/sass/index.js';
 import {
   codemodOptions,
   options,
-} from '../../../../helpers/shared-test-setups/ember-app/typescript.js';
+} from '../../../../helpers/shared-test-setups/ember-app/sass.js';
 
-test('migration | ember-app | steps | create-options > typescript', function () {
+test('migration | ember-app | steps | create-options > sass', function () {
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(createOptions(codemodOptions), options);
