+++ tests/migration/ember-engine/steps/create-options/sass.test.ts
@@ -1,13 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
 import { createOptions } from '../../../../../src/migration/ember-engine/steps/index.js';
-import { inputProject } from '../../../../fixtures/ember-engine/javascript/index.js';
+import { inputProject } from '../../../../fixtures/ember-engine/sass/index.js';
 import {
   codemodOptions,
   options,
-} from '../../../../helpers/shared-test-setups/ember-engine/javascript.js';
+} from '../../../../helpers/shared-test-setups/ember-engine/sass.js';
 
-test('migration | ember-engine | steps | create-options > javascript', function () {
+test('migration | ember-engine | steps | create-options > sass', function () {
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(createOptions(codemodOptions), options);
